ML Estimators for a Structured Correlation Matrix

Recursively computes the the correlation matrix and the degrees of freedom by Maximum-Likelihood (ML) of a t-copula with isotropic structure.

StrucTMLE(X, K, Tolerance = 10^(-10))

Arguments

X
A matrix made of timeseries invariants.

K
A numeric scalar with the number of factors.

Tolerance
A numeric scalar used for convergence. By default, the algorithm stops it's computations when the error is smaller than 10^(-10).

Value

A list with 2 components: Nu (the degrees of freedom) and C (the correlation matrix).

References

Meucci, Attilio, Estimation of Structured T-Copulas (April 2008). Available at SSRN: https://ssrn.com/abstract=1126401 or http://dx.doi.org/10.2139/ssrn.1126401.

Attilio Meucci (2020). https://www.mathworks.com/matlabcentral/fileexchange/19751-estimation-of-structured-t-copulas, MATLAB Central File Exchange. Retrieved October 14, 2020.

Examples

data("DB_SwapParRates")
Rates <- DB_SwapParRates

# The first difference is stationary
X <- Rates[2:nrow(Rates), ] - Rates[1:nrow(Rates) - 1, ]

# fit the model
StrucTMLE(X = X, K = 3)

#> $Nu
#> [1] 49
#> 
#> $C
#>            1y        2y        5y        7y       10y       15y       30y
#> 1y  1.0000000 0.9728346 0.8947780 0.8684387 0.8281359 0.8105573 0.7479613
#> 2y  0.9728346 1.0000000 0.9560076 0.9354872 0.9000657 0.8808805 0.8107507
#> 5y  0.8947780 0.9560076 1.0000000 0.9877988 0.9729065 0.9602568 0.9049002
#> 7y  0.8684387 0.9354872 0.9877988 1.0000000 0.9850406 0.9767635 0.9330718
#> 10y 0.8281359 0.9000657 0.9729065 0.9850406 1.0000000 0.9891340 0.9627638
#> 15y 0.8105573 0.8808805 0.9602568 0.9767635 0.9891340 1.0000000 0.9755292
#> 30y 0.7479613 0.8107507 0.9049002 0.9330718 0.9627638 0.9755292 1.0000000
#>